Scoping purgeCss to prod only, added explanation in comment.

+++ files/postcss.config.js
@@ -1,3 +1,4 @@
+const isProd = process.env.NODE_ENV === 'production'
 const purgeCss = require('@fullhuman/postcss-purgecss')
 const globAll = require('glob-all')
 const noComment = require('postcss-discard-comments')
@@ -9,10 +10,17 @@ const sorter = require('css-declaration-sorter')
 // const whitelister = require('purgecss-whitelister')
 
 
+/*
+  NOTE:
+  We need to scope `purgeCss` to production.
+  During development, CSS assets may have been previously purged,
+  leaving incomplete classes to play with when changing your JavaScript.
+  You might try to add classes in JS and see no change. This avoids that.
+*/
 module.exports = {
   plugins: [
     // https://goo.gl/igXRk6 - explains why we're using purge-css here and not as a Webpack plugin.
-    purgeCss({
+    isProd && purgeCss({
       // Optionally whitelist 3rd party libraries:
       // whitelist: whitelister('./node_modules/some-library/styles.css'),
       content: globAll.sync([
@@ -28,5 +36,5 @@ module.exports = {
     autoprefixer(),
     colormin(),
     sorter()
-  ]
+  ].filter(Boolean)
 }
